backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
python-dotenv
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pandas
scikit-learn==1.7.2
google-genai
dnspython